generate note
